+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Add `sdwan_banner_feature_template` resource and data source
 - Add `sdwan_tenant_feature_template` resource and data source
 - Add `sdwan_pim_feature_template` resource and data source
+- Add `sdwan_bridge_feature_template` resource and data source
 
 ## 0.3.6